+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [18.1.0]
+
+### Added
+
+- os
+  - bl64_os_match_compatible: check that the current OS is compatible against the target list
+  - bl64_os_check_version: added compatibility mode for OS version
+
 ## [18.0.0]
 
 ### Fixed